+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autons/Red01All.java
@@ -27,17 +27,17 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autons;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import org.firstinspires.ftc.robotcore.external.matrices.VectorF;
+import org.firstinspires.ftc.teamcode.AutoCommon;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import com.vuforia.CameraDevice;
-
 
 /**
  * This file illustrates the concept of driving a path based on encoder counts.
@@ -66,15 +66,15 @@ import com.vuforia.CameraDevice;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name="Blue 02 (All Actions)", group="OnBot")
+@Autonomous(name="RED 01 (All Actions)", group="OnBot")
 
 //@Disabled
-public class Blue02All extends LinearOpMode {
+public class Red01All extends LinearOpMode {
     private ElapsedTime     runtime = new ElapsedTime();
 
     AutoCommon auto=null;
 
-    Boolean blue=true;
+    Boolean blue=false;
 
     double driveSpeed = .3;
     double strafeSpeed = .3;
@@ -89,9 +89,13 @@ public class Blue02All extends LinearOpMode {
 
         waitForStart();
 
-        int pos = auto.getPos(true);
+        int pos = auto.getPos(false);
 
-        auto.encoderDrive(driveSpeed,500,10, false);
+        auto.encoderDrive(driveSpeed,-700,10, false);
+
+        auto.moveSpinner(spinnerSpeed, 2, true);
+
+        auto.encoderDrive(driveSpeed,1600,10, false);
 
         auto.encoderStrafe(strafeSpeed,10,1000,true,false,false);
 
@@ -99,9 +103,6 @@ public class Blue02All extends LinearOpMode {
 
         auto.encoderStrafe(strafeSpeed,10,1200,false,false,false);
 
-        auto.encoderDrive(driveSpeed,-2000,10, false);
-
-        auto.encoderStrafe(strafeSpeed,10,1200,true,false,false);
-
+        auto.encoderDrive(driveSpeed, 3000, 10,false);
     }
 }
