+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autons/Red02All.java
@@ -27,16 +27,18 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autons;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import org.firstinspires.ftc.robotcore.external.matrices.VectorF;
+import org.firstinspires.ftc.teamcode.AutoCommon;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import com.vuforia.CameraDevice;
+
 
 /**
  * This file illustrates the concept of driving a path based on encoder counts.
@@ -65,10 +67,10 @@ import com.vuforia.CameraDevice;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name="RED 01 (All Actions)", group="OnBot")
+@Autonomous(name="RED 02 (All Actions)", group="OnBot")
 
 //@Disabled
-public class Red01All extends LinearOpMode {
+public class Red02All extends LinearOpMode {
     private ElapsedTime     runtime = new ElapsedTime();
 
     AutoCommon auto=null;
@@ -90,11 +92,7 @@ public class Red01All extends LinearOpMode {
 
         int pos = auto.getPos(false);
 
-        auto.encoderDrive(driveSpeed,-700,10, false);
-
-        auto.moveSpinner(spinnerSpeed, 2, true);
-
-        auto.encoderDrive(driveSpeed,1600,10, false);
+        auto.encoderDrive(driveSpeed,-500,10, false);
 
         auto.encoderStrafe(strafeSpeed,10,1000,true,false,false);
 
@@ -102,6 +100,9 @@ public class Red01All extends LinearOpMode {
 
         auto.encoderStrafe(strafeSpeed,10,1200,false,false,false);
 
-        auto.encoderDrive(driveSpeed, 3000, 10,false);
+        auto.encoderDrive(driveSpeed,2000,10, false);
+
+        auto.encoderStrafe(strafeSpeed,10,1200,true,false,false);
+
     }
 }
